invalid read fix slash at end of line
